Ajax_Django/requirements.txt
Django==2.0.6
Pillow==5.2.0
pytz==2018.4
